bloc_2/src_code/scripts/requirements_airflow.txt
# Database connectors
pymongo
psycopg2-binary
snowflake-connector-python

# AWS
boto3
apache-airflow-providers-amazon

# Airflow providers
apache-airflow-providers-snowflake
apache-airflow-providers-postgres
apache-airflow-providers-mongo

# Data processing
pandasql
pandas

# Utils
python-dotenv